Fix typing short -> int

+++ pymartini/martini.pyx
@@ -127,7 +127,7 @@ cdef class Tile:
         cdef int k
         cdef unsigned short ax, ay, bx, by, mx, my, cx, cy
         cdef float interpolated_height, middle_error
-        cdef unsigned short middle_index, left_child_index, right_child_index
+        cdef unsigned int middle_index, left_child_index, right_child_index
 
         # iterate over all possible triangles, starting from the smallest level
         for i in range(self.max_num_triangles - 1, -1, -1):
